+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support for partially saved projects (KiCad 6). (See INTI-CMNB/kicad_auto#11)
 
 ## [1.6.13] - 2022-06-06
 ### Added